Serial : AT91 : Enable PDC and DMA RXBRK
The receiver detects a break condition when all data, parity and stop bits are low.
The PDC or DMA detect the RXBRK bit in US_CSR and enable RXBRK interrput, and meantime pdc or dma rx calculate the
break count.
A break condition will last for a specific duration, the receive data will be ignored during the duration.

Signed-off-by: Elen Song <elen.song@atmel.com>