mtd: atmel_nand: move the hsmc_clk from nfc node to nand node
Also add a new sama5d3_nand compatiable string for sama5d3 nand.

For sama5d3, sama5d4 chip, the pmecc become part of HSMC, they need the
HSMC clock enabled to work.
The NFC is a sub feature for current nand driver, it can be disabled.
But if HSMC clock is controlled by NFC, so disable NFC will also disable
the HSMC clock. then, it will make the PMECC fail to work.

So the solution is move the HSMC clock out of NFC to nand node. When
nand driver probed, it will check whether the chip has HSMC, if yes then
it will require a HSMC clock.

Change the device trees as well and use the newly introduced compatible string
for sama5d3. It's done in this patch to prevent configuration mismatch.

Signed-off-by: Josh Wu <josh.wu@atmel.com>
Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>
[nicolas.ferre@atmel.com: squash DT patches to prevent bisect issues]
Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>