ARM: at91/dt: sama5d2 Xplained: add sdmmc0 node
Add sdmmc0 node. Don't use card detect signal since it is inverted on
the board. We should have cd = 0 if the card is present. Hopefully, not
muxing the card detect signal will set it to 0 by default. Be careful,
there will be a metal fix setting it to 1 by default.

Signed-off-by: Ludovic Desroches <ludovic.desroches@atmel.com>