ARM: at91/dt: sama5d2: add M_CAN device nodes
Due to the CAN Message RAM can be shared by multi M_CAN instances,
so the base address of the CAN Message RAM for CAN0 and CAN1 are
shared at 0x00210000.

The default base address of the CAN Message RAM for CAN0 and CAN1
are configured at 0x200000 in CAN Memories Address-based Register.
To avoid conflict with SRAM map for pm, change them to 0x2100000
in the AT91Bootstrap.

As said in SAMA5D2 datasheet, the CAN clock is recommended to use
frequencies of 20, 40 or 80 MHz, and to select the UPLLCK(480 MHz)
as source clock. So, the "assigned-clock-rates" property has
three options: 20000000, 40000000, and 80000000. The
"assigned-clock-parents" property should be referred to utmi fixedly.

Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>