ARM: dts: at91: sama5d2 Xplained: Correct the macb irq pinctrl node
All pinctrl nodes for the Atmel pinctrl controller need to have their
bias configuration explicitly defined. Otherwise, the pinctrl mapping
is not valid.

It works for now as the pinctrl driver proceeds even with invalid
mappings, but this can become an issue, if the pinctrl driver starts
to require valid mappings. Additionally, the pin is not protected from
being remapped later by an other driver.

There is an external 1k pull-up to 3.3V, so no bias is required on
the Ethernet PHY's interrupt line.

Signed-off-by: Romain Izard <romain.izard.pro@gmail.com>
Acked-by: Ludovic Desroches <ludovic.desroches@atmel.com>
[nicolas.ferre@atmel.com: adapt to linux-4.1-at91 branch]
Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>