mtd: atmel-quadspi: remove buggy DMA support
The DMA support is buggy when we start using /dev/mtdblockX instead of
/dev/mtdX.

Indeed the mtd char device driver uses chunks of memory allocated with
kmalloc(), which guarantees that the returned pages are contiguous in
physical space, whereas the mtd block device driver uses vmalloc() for
its allocations. Memory regions returned by vmalloc() are not guaranteed
to be contiguous in physical space.

Also the DMA support relied on dma_map_single() which is not suited when
the memory buffer to be mapped is longer than a single page.

Finally, the API for DMA memcpy expects pointers to the source and
destination buffers instead of scatter-gather lists.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@atmel.com>