pinctrl: at91-pio4: fix pull-up/down logic
The default configuration of a pin is often with a value in the pull-up/down
field at chip reset. So, even if the internal logic of the controller prevents
writing a configuration with pull-up and pull-down at the same time, we must
ensure explicitly this condition before writing the register.

This was leading to a pull-down condition not taken into account for instance.

Signed-off-by: Ludovic Desroches <ludovic.desroches@atmel.com>